Add checkout functionality and integrate shipping service

+++ VirtualTryonWomenFashion/VirtualTryonWomenFashion.API/Controllers/CartController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using VirtualTryonWomenFashion.Service.DTO.Cart;
+using VirtualTryonWomenFashion.Service.Helpers;
 using VirtualTryonWomenFashion.Service.IServices;
 
 namespace VirtualTryonWomenFashion.API.Controllers;
@@ -21,11 +22,21 @@ public class CartController : ControllerBase
         try
         {
             var cartItems = await _cartService.GetCartItemsAsync();
-            return Ok(cartItems);
+            return Ok(new MessageModelWithData<object>()
+            {
+                StatusCode = StatusCodes.Status200OK,
+                Message = "Get cart successfully",
+                Data = cartItems
+            });
         }
         catch (Exception ex)
         {
-            return BadRequest(new { Message = ex.Message });
+            return BadRequest(new MessageModelWithData<object>()
+            {
+                StatusCode = StatusCodes.Status400BadRequest,
+                Message = ex.Message,
+                Data = null
+            });
         }
     }
     
@@ -35,43 +46,97 @@ public class CartController : ControllerBase
         try
         {
             var selectedCartItems = await _cartService.GetSelectedCartItemsAsync(productVariantIds);
-            return Ok(selectedCartItems);
+            return Ok(new MessageModelWithData<object>()
+            {
+                StatusCode = StatusCodes.Status200OK,
+                Message = "Get selected item in cart successfully",
+                Data = selectedCartItems
+            });
         }
         catch (Exception ex)
         {
-            return BadRequest(new { Message = ex.Message });
+            return BadRequest(new MessageModelWithData<object>()
+            {
+                StatusCode = StatusCodes.Status400BadRequest,
+                Message = ex.Message,
+                Data = null
+            });
         }
     }
     
     [HttpPost("add-item")]
     public async Task<IActionResult> AddItemToCartAsync([FromBody] RequestAddProductToCart requestAddProductToCart)
     {
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(new MessageModelWithData<object>()
+            {
+                StatusCode = StatusCodes.Status400BadRequest,
+                Message = " Invalid input data",
+                Data = null
+            });
+        }
         try
         {
+            
             bool result = await _cartService.AddProductToCartAsync(requestAddProductToCart);
             if (result)
             {
-                return Ok(new { Message = "Product added to cart successfully." });
+                return Ok(new MessageModelWithData<object>()
+                {
+                    StatusCode = StatusCodes.Status201Created,
+                    Message =  "Product added to cart successfully." ,
+                    Data = null
+                });
             }
-            return BadRequest(new { Message = "Failed to add product to cart." });
+            return BadRequest(new MessageModelWithData<object>()
+            {
+                StatusCode = StatusCodes.Status400BadRequest,
+                Message =  "Failed to add product to cart." ,
+                Data = null
+            });
         }
         catch (Exception ex)
         {
-            return BadRequest(new { Message = ex.Message });
+            return BadRequest(new MessageModelWithData<object>()
+            {
+                StatusCode = StatusCodes.Status400BadRequest,
+                Message = ex.Message,
+                Data = null
+            });
         }
     }
     
     [HttpPut("update-quantity")]
     public async Task<IActionResult> UpdateProductQuantityAsync([FromBody] RequestAddProductToCart requestUpdateProductQuantity)
     {
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(new MessageModelWithData<object>()
+            {
+                StatusCode = StatusCodes.Status400BadRequest,
+                Message = " Invalid input data",
+                Data = null
+            });
+        }
         try
         {
             bool result = await _cartService.UpdateProductQuantityAsync(requestUpdateProductQuantity);
             if (result)
             {
-                return Ok(new { Message = "Product quantity updated successfully." });
+                return Ok(new MessageModelWithData<object>()
+                {
+                    StatusCode = StatusCodes.Status200OK,
+                    Message =  "Product quantity updated successfully." ,
+                    Data = null
+                });
             }
-            return BadRequest(new { Message = "Failed to update product quantity." });
+            return BadRequest(new MessageModelWithData<object>()
+            {
+                StatusCode = StatusCodes.Status400BadRequest,
+                Message =  "Failed to update product quantity." ,
+                Data = null
+            });
         }
         catch (Exception ex)
         {
@@ -82,18 +147,66 @@ public class CartController : ControllerBase
     [HttpDelete("remove-item/{productVariantId}")]
     public async Task<IActionResult> RemoveItemFromCartAsync(string productVariantId)
     {
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(new MessageModelWithData<object>()
+            {
+                StatusCode = StatusCodes.Status400BadRequest,
+                Message = " Invalid input data",
+                Data = null
+            });
+        }
         try
         {
             bool result = await _cartService.RemoveProductFromCartAsync(productVariantId);
             if (result)
             {
-                return Ok(new { Message = "Product removed from cart successfully." });
+                return Ok(new MessageModelWithData<object>()
+                {
+                    StatusCode = StatusCodes.Status200OK,
+                    Message =  "Product removed from cart successfully.",
+                    Data = null
+                });
             }
-            return BadRequest(new { Message = "Failed to remove product from cart." });
+            return BadRequest(new MessageModelWithData<object>()
+            {
+                StatusCode = StatusCodes.Status400BadRequest,
+                Message =  "Failed to remove product from cart.",
+                Data = null
+            });
         }
         catch (Exception ex)
         {
-            return BadRequest(new { Message = ex.Message });
+            return BadRequest(new MessageModelWithData<object>()
+            {
+                StatusCode = StatusCodes.Status400BadRequest,
+                Message = ex.Message,
+                Data = null
+            });
+        }
+    }
+    
+    [HttpPost("checkout")]
+    public async Task<IActionResult> CheckoutAsync([FromBody] RequestCheckout requestCheckout)
+    {
+        try
+        {
+            var checkoutDetails = await _cartService.CheckoutAsync(requestCheckout);
+            return Ok(new MessageModelWithData<object>()
+            {
+                StatusCode = StatusCodes.Status200OK,
+                Message = "Checkout successfully",
+                Data = checkoutDetails
+            });
+        }
+        catch (Exception ex)
+        {
+            return BadRequest(new MessageModelWithData<object>()
+            {
+                StatusCode = StatusCodes.Status400BadRequest,
+                Message = ex.Message,
+                Data = null
+            });
         }
     }
